backend/requirements.txt
fastapi==0.115.2 
uvicorn==0.32.0 
pandas==2.2.3 
openai==1.51.2 
python-dotenv==1.0.1 
pydantic==2.9.2 
openpyxl==3.1.5